+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Docker formula
 ==============
 
+2016-04-10 - 0.3.0
+ - Removed support for Ubuntu 15.04 (repos no longer supported by Docker)
+ - Added Docker Volume Support for local-persist driver.
+
 2016-04-03 - 0.2.0
  - Major refactoring of formula.
  - Added commercial engine support.